% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPN.R
\name{MPN}
\alias{MPN}
\title{Most Probable Number (MPN) calculator}
\usage{
MPN(g, m, t)
}
\arguments{
\item{g}{the number of positive (or growth) tubes in the jth dilution}

\item{m}{the amount of the original sample put in the tubes at each dilution.}

\item{m}{the number of tubes at each dilution}
}
\value{
The most probable  \code{x} and \code{y}.
}
\description{
Most Probable Number (MPN) calculator
}
\examples{
MPN(g=c(5,3,0),m=c(0.1,0.01,0.001),t=10)
}

